# app/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenMark API",
    "description": "API for PDF document viewing with annotation capabilities. OpenMark provides secure access to PDF documents with support for notes, highlights, and collaborative annotations.\n\n## Authentication\n\nOpenMark uses two types of tokens:\n\n1. **Auth Token**: Standard JWT token obtained via `/api/authenticate`. Used for most API calls.\n2. **Document Access Token (DAT)**: Self-contained JWT for document access. Survives page refresh.\n\n## Quick Start\n\n### Option 1: Quick View (Single Call)\n```\nPOST /api/quickView\n{\n  \"username\": \"user\",\n  \"password\": \"pass\",\n  \"documentId\": \"sample\"\n}\n```\n\n### Option 2: Three-Step Process\n1. Authenticate → Get token\n2. Request document → Get DAT\n3. View document with DAT",
    "version": "1.0.0",
    "contact": {
      "name": "OpenMark Support"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current server"
    }
  ],
...
  "paths": {
...
    "/viewDocument": {
      "get": {
        "tags": ["Documents"],
        "summary": "View document in PDF viewer",
        "description": "Redirect to the PDF viewer with the specified document. Use the Document Access Token (DAT) for authentication.",
        "operationId": "viewDocument",
        "parameters": [
          {
            "name": "dat",
            "in": "query",
            "required": true,
            "description": "Document Access Token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to PDF viewer"
          },
          "401": {
            "description": "Invalid or expired DAT"
          }
        }
      }
    },
...
    "/statistics": {
      "get": {
        "tags": ["User Data"],
        "summary": "Get user statistics",
        "description": "Retrieve usage statistics for the authenticated user including documents viewed, notes created, and highlights created.",
        "operationId": "getStatistics",
        "security": [
          {"bearerAuth": []}
        ],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/history": {
      "get": {
        "tags": ["User Data"],
        "summary": "Get consultation history",
        "description": "Retrieve document consultation history for the authenticated user with pagination support.",
        "operationId": "getHistory",
        "security": [
          {"bearerAuth": []}
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of records to return",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Number of records to skip",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "History retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /authenticate endpoint"
      },
      "datAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "dat",
        "description": "Document Access Token for document-specific operations"
      }
    },
    "schemas": {
...
      "StatisticsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statistics": {
            "type": "object",
            "properties": {
              "documents_viewed": {
                "type": "integer",
                "description": "Number of documents viewed"
              },
              "notes_created": {
                "type": "integer",
                "description": "Number of notes created"
              },
              "highlights_created": {
                "type": "integer",
                "description": "Number of highlights created"
              },
              "last_activity": {
                "type": "string",
                "format": "date-time",
                "description": "Last activity timestamp"
              }
            }
          }
        }
      },
      "HistoryResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoryEntry"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total number of history entries"
          },
          "limit": {
            "type": "integer",
            "description": "Applied limit"
          },
          "offset": {
            "type": "integer",
            "description": "Applied offset"
          }
        }
      },
      "HistoryEntry": {
        "type": "object",
        "properties": {
          "document_id": {
            "type": "string",
            "description": "Document identifier"
          },
          "document_name": {
            "type": "string",
            "description": "Document display name"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "View timestamp"
          },
          "ip_address": {
            "type": "string",
            "description": "Client IP address"
          },
          "duration_seconds": {
            "type": "integer",
            "description": "Viewing duration in seconds"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      }
    }
  }
}